# docs/swagger.yaml
definitions:
  handlers.BasicErrorResponse:
    properties:
      error:
        default: An error occurred
        type: string
    type: object
  handlers.BasicSuccessResponse:
    properties:
      message:
        default: Operation successful
        type: string
    type: object
  handlers.BatchUpdateScheduleRequest:
    properties:
      arrangements:
        items:
          $ref: '#/definitions/models.ScheduleEntry'
        type: array
      date:
        type: string
    type: object
  handlers.CreateCardRequest:
    properties:
      name:
        type: string
    type: object
  handlers.CreateEntryRequest:
    properties:
      schedule_date:
        default: YYYY-MM-DD
        description: Expecting YYYY-MM-DD
        type: string
      start_time:
        default: HH:MM:SS
        description: Expecting HH:MM:SS
        type: string
      task_description:
        default: task description
        description: Expecting "task description"
        type: string
    required:
    - schedule_date
    - start_time
    - task_description
    type: object
  handlers.DataImportRequest:
    properties:
      blocks:
        items:
          $ref: '#/definitions/models.Block'
        type: array
      cards:
        items:
          $ref: '#/definitions/models.Card'
        type: array
    type: object
  models.Block:
    properties:
      card_id:
        description: Optional association to a card, can be null
        type: string
      coordinates:
        description: 'For Polygon and LineString: [[lon, lat], [lon, lat], ...]'
        items:
          items:
            format: float64
            type: number
          type: array
        type: array
      geometry_type:
        description: '"Polygon" or "LineString"'
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.Card:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  models.Conductor:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  models.CreateBlockRequest:
    properties:
      card_id:
        type: string
      coordinates:
        description: Expecting array of [lon, lat] pairs
        items:
          items:
            format: float64
            type: number
          type: array
        type: array
      geometry_type:
        type: string
      name:
        type: string
    required:
    - coordinates
    - geometry_type
    - name
    type: object
  models.MapBlocksResponse:
    properties:
      blocks_count:
        type: integer
      polygons:
        description: 'Nested array: [Polygon][Point][Lat/Lon]'
        items:
          items:
            items:
              format: float64
              type: number
            type: array
          type: array
        type: array
    type: object
  models.ScheduleEntry:
    properties:
      card_id:
        type: string
      conductor_name:
        type: string
      id:
        type: string
      location:
        type: string
      schedule_date:
        type: string
      start_time:
        type: string
      task_description:
        type: string
    type: object
  models.Venue:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /conductors:
    get:
      consumes:
      - application/json
      description: Get a list of all conductors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Conductor'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: List conductors
      tags:
      - conductors
  /maps/blocks:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Block'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: List user-defined blocks
      tags:
      - map
    post:
      consumes:
      - application/json
      parameters:
      - description: Block creation request
        in: body
        name: block
        required: true
        schema:
          $ref: '#/definitions/models.CreateBlockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Block'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: Create a new user-defined block
      tags:
      - map
  /maps/blocks/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Block ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BasicSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: Delete a user-defined block
      tags:
      - map
  /maps/blocks/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file with block data
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BasicSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: Upload a CSV file to sync blocks
      tags:
      - map
  /maps/cards:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Card'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: List user-defined cards (placeholder)
      tags:
      - map
    post:
      consumes:
      - application/json
      parameters:
      - description: Card creation request
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BasicSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: Create a new user-defined card
      tags:
      - map
  /maps/cards/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BasicSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: Delete a user-defined card
      tags:
      - map
  /maps/cards/{id}/blocks:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Card Name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BasicSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: Delete all user-defined blocks associated with a card
      tags:
      - map
  /maps/generated-blocks:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MapBlocksResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: List generated blocks in Lekki Phase 1
      tags:
      - map
  /maps/import:
    post:
      consumes:
      - application/json
      description: Accepts a JSON payload containing arrays of cards and blocks to
        restore local database states.
      parameters:
      - description: Data import request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.DataImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BasicSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: Import data into the database
      tags:
      - map
  /schedules:
    get:
      consumes:
      - application/json
      description: Get a list of schedule entries for a given date range
      parameters:
      - description: Start date in YYYY-MM-DD format
        in: query
        name: start_date
        required: true
        type: string
      - description: End date in YYYY-MM-DD format
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ScheduleEntry'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: List schedule entries
      tags:
      - schedules
    post:
      consumes:
      - application/json
      description: Create a schedule entry for a specified date
      parameters:
      - description: Schedule entry details
        in: body
        name: default
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateEntryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ScheduleEntry'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: Create a schedule entry
      tags:
      - schedules
    put:
      consumes:
      - application/json
      description: Update multiple schedule entries in a single request
      parameters:
      - description: Schedule entry details
        in: body
        name: default
        required: true
        schema:
          $ref: '#/definitions/handlers.BatchUpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BasicSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: Batch Update a schedule entry
      tags:
      - schedules
  /schedules/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a PDF file to sync schedule entries. Optionally specify
        a year to filter the entries.
      parameters:
      - description: PDF file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Optional year to filter entries (e.g., 2026)
        in: formData
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BasicSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: Upload a schedule PDF
      tags:
      - schedules
  /venues:
    get:
      consumes:
      - application/json
      description: retrieves a list of all available venues and responds with a JSON-encoded
        array of venue objects.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Venue'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.BasicErrorResponse'
      summary: List venues
      tags:
      - venue
swagger: "2.0"